feat: Add dual-format CSV output and reorganize project structure
## Major Features Added

### CSV Formatting System
- Add csv_formatter.py: Convert between long/wide format CSV outputs
- Long format: Each row = location-year combination (existing)
- Wide format: Each row = location, columns = index_year
- Historical data filtering (1950-2014) excludes projections
- Support for all 12 climate indices across multiple years

### Project Structure Reorganization
- scripts/: Processing and analysis tools (8 files)
- data/: Organized test data and samples (7 test CSV files)
- docs/: Documentation and reports
- benchmarks/: Performance data and visualizations
- tests/: Testing scripts and utilities

### Documentation Enhancement
- Add CLAUDE.md: Comprehensive project architecture guide
- Add CLEANUP_SUMMARY.md: Detailed reorganization summary
- Update README.md: Reflect new structure and CSV capabilities
- Enhanced configuration with CSV format options

### Technical Improvements
- Vectorized point extraction for large datasets
- Memory-efficient processing with pandas pivot operations
- Configurable year ranges for historical data focus
- Command-line interface for batch CSV processing

## Files Changed
- 24 files moved/renamed to organized directories
- 6 new files added (formatters, docs, examples)
- Root directory reduced from 32 to 14 items
- All Python cache directories cleaned

## Benefits
- Clean separation of core pipeline vs utilities
- Dual CSV formats support different analysis workflows
- Historical data focus avoids projection confusion
- Professional project organization for maintenance
