+++ CLEANUP_SUMMARY.md
@@ -1,0 +1,80 @@
+# Root Directory Cleanup Summary
+
+## Before Cleanup
+The root directory contained **32 files and directories** with mixed content types scattered throughout.
+
+## Organization Structure Created
+
+### 📁 **scripts/** - Processing & Analysis Tools (8 files)
+- `csv_formatter.py` - CSV format converter (long ↔ wide)
+- `efficient_extraction.py` - Optimized point extraction
+- `fast_point_extraction.py` - Alternative extraction method
+- `point_extraction.py` - Legacy extraction script
+- `climate_indices_corrections.py` - Climate indices corrections
+- `format_csv_example.py` - CSV formatting example
+- `test_annual_temp.py` - Annual temperature testing
+- `visualize_temp.py` - Results visualization
+
+### 📁 **data/** - Data Files
+#### **test_data/** (7 files)
+- `parcel_coordinates.csv` - Main parcel coordinates (641KB)
+- `parcel_coordinates_sample.csv` - Sample coordinates
+- `parcel_test_1000.csv` - 1000-parcel test subset
+- `precip_test_2010.csv` - Precipitation test data (188KB)
+- `real_data_full_2010.csv` - Full 2010 real data (9.2MB)
+- `real_data_test_2010.csv` - 2010 test data (378KB)
+- `test_enhanced_output.csv` - Enhanced output test (62KB)
+
+#### **sample_data/**
+- Sample files moved from previous sample_data directory
+
+### 📁 **docs/** - Documentation (2 files)
+- `MERGE_INSTRUCTIONS.md` - Development merge instructions
+- `SCIENTIFIC_REVIEW_REPORT.md` - Scientific review and validation
+
+### 📁 **benchmarks/** - Performance Data (2 files)
+- `benchmark_results.json` - Performance metrics
+- `benchmark_results.png` - Benchmark visualization (68KB)
+
+### 📁 **tests/** - Testing Scripts (1 file)
+- `run_test.sh` - Test execution script
+
+## Root Directory After Cleanup
+
+### **Core Files Remaining (11 items):**
+```
+├── src/                    # Core pipeline modules
+├── config_simple.yaml     # Configuration template
+├── requirements.txt        # Python dependencies
+├── setup_env.sh           # Environment setup
+├── README.md              # Main documentation
+├── CLAUDE.md              # Comprehensive project guide
+├── outputs/               # Processing results
+├── logs/                  # Processing logs
+├── .venv/                 # Python virtual environment
+├── .git/                  # Git repository
+└── .gitignore             # Git ignore rules
+```
+
+### **Removed Items:**
+- `__pycache__/` - Python cache directory (removed)
+- `.pytest_cache/` - Pytest cache directory (removed)
+- All scattered CSV test files (organized into data/)
+- All Python scripts (organized into scripts/)
+- Documentation files (organized into docs/)
+
+## Benefits of Cleanup
+
+1. **Clear Separation**: Core pipeline (`src/`) vs utilities (`scripts/`) vs data (`data/`)
+2. **Easier Navigation**: Related files grouped logically
+3. **Development Focus**: Root directory shows only essential project files
+4. **Maintenance**: Easier to find and update specific file types
+5. **Documentation**: Clear project structure for new contributors
+
+## Updated File Counts
+- **Before**: 32 items in root directory
+- **After**: 11 items in root directory
+- **Files Organized**: 19 files moved to appropriate subdirectories
+- **Cache Cleaned**: 2 temporary directories removed
+
+The project now follows a clean, logical structure that separates concerns and makes navigation much easier for both development and usage.
