+++ benchmarks/benchmark_results.json
@@ -1,0 +1,48 @@
+{
+  "efficient": {
+    "times": [
+      2.7609421529923566,
+      0.10533268001745455,
+      0.20401683801901527,
+      0.3893096490064636,
+      0.5585943140031304
+    ],
+    "memory": [
+      31.650522232055664,
+      28.335773468017578,
+      49.87157440185547,
+      92.94627666473389,
+      136.0208158493042
+    ],
+    "counts": [
+      100,
+      500,
+      1000,
+      2000,
+      3000
+    ]
+  },
+  "fast": {
+    "times": [
+      0.3825963480048813,
+      1.7480803779908456,
+      3.4102364360005595,
+      6.72745906401542,
+      10.100490734010236
+    ],
+    "memory": [
+      1.2828540802001953,
+      5.453927993774414,
+      10.42220401763916,
+      20.66323471069336,
+      30.907198905944824
+    ],
+    "counts": [
+      100,
+      500,
+      1000,
+      2000,
+      3000
+    ]
+  }
+}
