+++ data/test_data/parcel_coordinates_sample.csv
@@ -1,0 +1,101 @@
+saleid,parcelid,parcel_level_latitude,parcel_level_longitude
+600007,1,39.7964,-121.445
+600008,1,39.7561,-121.474
+600009,1,39.7596,-121.519
+600010,1,39.8234,-121.574
+600011,1,39.6983,-121.193
+600012,1,39.9935,-121.515
+600013,1,39.8389,-121.428
+600014,1,39.7518,-121.494
+600015,1,39.6148,-121.408
+600016,1,39.7585,-121.5
+600017,1,39.5679,-121.239
+600018,1,39.7536,-121.482
+600019,1,39.7574,-121.526
+600020,1,39.8756,-121.621
+600021,1,39.5547,-121.232
+600022,1,39.6731,-121.314
+600023,1,39.5239,-121.176
+600024,1,39.5199,-121.251
+600025,1,39.6381,-121.305
+600026,1,39.6803,-121.436
+600027,1,39.5215,-121.178
+600028,1,39.5995,-121.217
+600029,1,39.9537,-121.66
+600032,1,38.2576,-120.429
+600033,1,38.2424,-120.315
+600034,1,38.4009,-120.436
+600035,1,38.4174,-120.485
+600045,1,41.6774,-123.893
+600046,1,41.8394,-124.126
+600047,1,41.6782,-123.9
+600048,1,41.5263,-124.032
+600049,1,41.7135,-123.854
+600050,1,41.9284,-124.071
+600051,1,41.8426,-124.004
+600052,1,41.8499,-124.088
+600053,1,41.9704,-123.828
+600054,1,41.9125,-124.066
+600055,1,41.9838,-123.817
+600056,1,41.8504,-124.149
+600057,1,41.7755,-123.948
+600058,1,41.9571,-124.186
+600059,1,41.848,-123.826
+600060,1,41.8578,-123.796
+600061,1,41.8374,-123.814
+600062,1,41.7637,-124.173
+600063,1,41.8577,-123.8
+600064,1,41.8662,-124.119
+600065,1,41.8603,-124.078
+600066,1,41.981,-123.962
+600067,1,41.9791,-123.808
+600068,1,41.7267,-123.975
+600069,1,41.8713,-124.079
+600070,1,41.8774,-124.084
+600071,1,41.8668,-124.074
+600072,1,41.7484,-123.992
+600073,1,41.8325,-124.149
+600074,1,41.8677,-124.062
+600075,1,41.8689,-124.083
+600076,1,41.8684,-124.075
+600077,1,41.8602,-124.069
+600078,1,41.846,-124.149
+600079,1,41.7308,-123.979
+600080,1,41.8451,-124.065
+600081,1,41.8432,-124.112
+600082,1,41.7063,-123.884
+600083,1,41.4675,-123.952
+600084,1,41.872,-124.064
+600085,1,41.8579,-123.793
+600086,1,41.7696,-124.166
+600102,1,38.6484,-120.612
+600103,1,38.5763,-120.458
+600104,1,38.9235,-120.904
+600105,1,38.8965,-120.563
+600106,1,38.7781,-120.768
+600107,1,38.9606,-120.478
+600108,1,38.971,-120.373
+600109,1,38.5331,-120.43
+600110,1,38.9056,-120.599
+600111,1,38.6511,-120.519
+600112,1,39.0069,-120.318
+600113,1,38.679,-120.567
+600114,1,38.8021,-120.596
+600115,1,38.5721,-120.422
+600116,1,38.9262,-120.38
+600117,1,38.6312,-120.631
+600118,1,38.907,-120.541
+600119,1,38.6265,-120.402
+600120,1,38.6509,-120.21
+600121,1,38.5371,-120.43
+600122,1,38.7707,-120.446
+600123,1,38.6961,-120.664
+600124,1,38.9309,-120.845
+600125,1,38.9068,-120.534
+600126,1,38.6173,-120.585
+600127,1,38.8045,-120.226
+600128,1,38.8158,-120.667
+600129,1,38.6529,-120.517
+600130,1,38.9071,-120.545
+600131,1,38.6228,-120.578
+600132,1,38.6683,-120.218
